test(vapor): split slot-outlet tests into submodules to satisfy file-length gate
