+++ crates/vize_atelier_vapor/src/tests.rs
@@ -799,78 +799,6 @@ fn test_compile_slot_outlet_preserves_dynamic_name() {
 }
 
 #[test]
-fn test_compile_bare_default_slot_outlet() {
-    let allocator = Bump::new();
-    let result = compile_vapor(&allocator, r#"<slot />"#, Default::default());
-
-    assert!(
-        result.error_messages.is_empty(),
-        "Expected no errors: {:?}",
-        result.error_messages
-    );
-
-    let code = normalize_code(&result.code);
-    assert_parses_as_module(&code);
-    insta::assert_snapshot!(code.as_str());
-}
-
-#[test]
-fn test_compile_named_slot_outlet_without_props() {
-    let allocator = Bump::new();
-    let result = compile_vapor(&allocator, r#"<slot name="head" />"#, Default::default());
-
-    assert!(
-        result.error_messages.is_empty(),
-        "Expected no errors: {:?}",
-        result.error_messages
-    );
-
-    let code = normalize_code(&result.code);
-    assert_parses_as_module(&code);
-    insta::assert_snapshot!(code.as_str());
-}
-
-#[test]
-fn test_compile_default_slot_outlet_with_fallback_only() {
-    let allocator = Bump::new();
-    let result = compile_vapor(
-        &allocator,
-        r#"<slot><div>fb</div></slot>"#,
-        Default::default(),
-    );
-
-    assert!(
-        result.error_messages.is_empty(),
-        "Expected no errors: {:?}",
-        result.error_messages
-    );
-
-    let code = normalize_code(&result.code);
-    assert_parses_as_module(&code);
-    insta::assert_snapshot!(code.as_str());
-}
-
-#[test]
-fn test_compile_slot_outlet_static_and_spread_props() {
-    let allocator = Bump::new();
-    let result = compile_vapor(
-        &allocator,
-        r#"<slot kind="primary" :row="item" v-bind="extra" />"#,
-        Default::default(),
-    );
-
-    assert!(
-        result.error_messages.is_empty(),
-        "Expected no errors: {:?}",
-        result.error_messages
-    );
-
-    let code = normalize_code(&result.code);
-    assert_parses_as_module(&code);
-    insta::assert_snapshot!(code.as_str());
-}
-
-#[test]
 fn test_compile_custom_directive_preserves_payloads() {
     let allocator = Bump::new();
     let result = compile_vapor(
